Upgrade esm to version 3.2.0
Unfortunately we were broken by a change in version 3.2.0 of esm. It no
longer exports a default export as the default key on the exports object
but rather moves it to be the actually exported thing. In commonjs
module terms they changed it such that...

```js
export default function myFunction() {}
```

... would come out as ...

```js
module.exports = function myFunction() {}
```

... rather than ...

```js
exports.default = function myFunction() {}
```